Replace git diff modal with inline diff view in chat panel (#146)
Replaces the full-screen GitDiffModal with an inline diff viewer that
renders in the same slot as FileViewer, keeping the Modified file list
sidebar visible for navigation. Clicking a file in "Modified" opens it
in diff mode; clicking in "All" opens source mode. A Source/Diff toggle
in the new FileContentToolbar allows switching views on any modified file.

- Add fileViewMode to useTabs with workspace snapshot persistence
- Create InlineDiffViewer (single-file diff with line comments)
- Create FileContentToolbar (path display, mode toggle, paste-to-prompt)
- Add ChatInput.appendText via forwardRef for comment pasting
- Highlight active file in ModifiedFileList
- Show diff icon in ConversationTabs when in diff mode
- Delete GitDiffModal and its tests
